Add a CI job to run unit tests with a minimal (non-dev) installation (#118)
Adds an additional CI job to run the unit tests with a minimal
configuration (without `--with dev`). This will ensure that the minimal
install of pixeltable continues to work.

This PR also reclassifies a few dependencies as non-dev (that are core
to pixeltable) and adds `skip_test_if_not_installed` to various tests
that depend on optional packages.