Partial fix for duplicate progress bars (temporary hack) (#231)
This is a temporary solution to fix the duplicate progress bars on
insertions. The duplicates were introduced when indexable strings were
added to Pixeltable via system/anonymous computed columns, resulting in
a frivolous progress bar showing up on every insertion and crossing
wires with the insertion console messages.

This is not a complete or long-term solution and is intended to hold us
over until we revamp progress bars more generally (which is on the
roadmap for open source v1).