Ensure FunctionCall return types have correct nullability (#347)
Ensures that a `FunctionCall` will have nullable return type if any of
its non-nullable parameters have a nullable input. This conforms to
Pixeltable's behavior of defaulting to `None` any time a `FunctionCall`
sees such an input.

Previously, the column type was reported as non-nullable but `None`s
were being generated, an inconsistency.