PXT-274 - Use GPUs for sentence_transformers when available
This commit changes the following sentence_transformers udf
- sentence_transformer
- sentence_transformer_list
- cross_encoder
- cross_encoder_list
to use the GPU when available by specifying the torch.device.

Testing: make test ran clean on laptop (assume, it uses mps gpu).
CI should run with cpu. Pending test in colab for cuda gpu.
It is not yet clear how to verify/confirm it uses the device.