PXT-274 - Pass torch.device to sentence_transformers (#361)
This commit changes the following sentence_transformers udf
- sentence_transformer
- sentence_transformer_list
- cross_encoder
- cross_encoder_list

to explicitly specify the torch.device. It changes the model
lookup to use the device at the time of the model creation.

Testing: make test and make typecheck ran clean on laptop (uses mps gpu).
Tested in colab for cuda gpu and cpu.