PXT-290: Allow APIs to accept table/column handles
This commit allows the drop_table API to accept a
table handle as an alternative to the table name.
It also adds an assertion in the catalog.Table._drop
and catalog.View._drop internal API to assert that
the dependencies are dropped by the time they are
called.

Testing: new tests that operated on their names