PXT-330 Fix a bug in create_insert_plan (#372)
This commit fixes a bug in Planner.create_insert_plan where an insert
continues to include a column even after it was dropped. This was
because it was using catalog.TableVersion.cols to get the list of
columns, but that tracks history of all the columns in the table/view.
Fix is to use catalog.TableVersion.cols_by_id that tracks the current
list of columns.

Testing: new testcase that reproduces the bug passes with the fix.