Partial bindings for `Function` (#363)
Introduces a `Function.using()` method that creates a partial binding of
the function parameters as an `ExprTemplateFunction`. If `f` is an n-ary
`Function`, then

`f.using(a=..., b=..., ...)`

is an (n-k)-ary function whose parameters are all the parameters of `f`
not among `a, b, ...`.

This is useful for specifying indices without having to create explicit
`@expr_udf`s.