PXT-352 gracefully error when using a catalog.Table after it is dropped (#383)
This commit changes the catalog.Table APIs to consistently check when
the table is already dropped, and error out gracefully and early. Before
some APIs did the check, few simply succeeded as if the table was still
active, and many failed with unfriendly errors errors during SQL
execution later.

Testing: added a new testcase to verify.