PXT-354: Bug: Indexed column for a view may not load correctly from persistence
Test repro for the bug and some experiments to figure the RCA.
When an embedding is created on a column, the `indexed_col_id` field of the
index metadata tracks the column on which this embedding was created.
For a view that can be ambiguous: it may be id of a column in the view itself,
or id of a column of any base table(s) that the view is derived from.
Every table/view generates its own “column id” independently.
But only the “column id” is persisted in the index metadatai, and so on
reloading it from persistence, it can be incorrectly interpreted for view.
The code in `TableVerison._init_schema()→_init_idxs()`  assumes the
`indexed_col_id` for an index is from the view’s columns.