PXT-354: Bug: Indexed column for a view may not load correctly from persistence
Added converters to upgrade persistence to include the indexed
column's table id in the index metadata, if it is missing.
Since we cannot know the correct value of the new field for past
views, upgrade keeps the old [buggy] behavior by assigning it
the table id of the view on which the index is created.

Testing: Before, test_db_migration would fail to load an
older database dump due to the missing field in the index
metadata. With this commit, it passes.