PXT-235 - do not accept bool literals in int columns
bool is a subclass of int and so when True or False is
passed as a literal for insert into a column of type int,
we dont raise any error. But in persistence the int
column is stored as a Bigint column and that does not
accept True or False literals. So the insert eventually
fails with an unfriendly error from the database later.
This commit catches such bool literals early and raises
a more user-friendly error before doing any database
operations.

Alternatively, if we wanted the flexibility to accept
these bool literals for an int column, we could fix
this by converting them to 1 and 0 before passing it
to persistence.

Testing: added new testcase to cover this code path.
Also, added test coverage for bool column bugs from
past (tracked under PXT-204, PXT-207) that are no
longer reproducible.