PXT-179 Add a if_exists parameter to create/add APIs
This commit adds a new parameter - if_exists - for the
create_dir API which can be used to handle scenarios
where a directory may already exist. It can direct
the API to not create the directory and return, or
raise it as an error, or replace the existing directory
with a new one.

Testing: improved existing testcases to test with
the new parameters.