PXT-179 Add a if_exists parameter to create/add APIs
This commit adds a new parameter - if_exists - for the
create_snapshot API which can be used to handle scenarios
where a table may already exist. It can direct the API
to not create the view and return the existing one, or
raise it as an error, or replace the existing view
with a new one. Also, consolidated common code.

Testing: added new testcases for the new parameter.