PXT-235 - do not accept bool literals in int columns (#400)
bool is a subclass of int and so when True or False is passed as a
literal to a column of type int, we dont raise an error. But in
persistence the int column is stored as Bigint and it does not accept
True or False literals. So the query fails with a cryptic insert error
from the database later. This commit catches bool literals early and
raises an error before doing any database operations.

Alternatively, if we wanted the flexibility to accept these bool
literals for an int column, we could fix this by converting them to 1
and 0 before doing any database operation.

Testing: added new testcase to cover this code path. 
Also, added test coverage for bool column bugs from past (tracked under
PXT-204, PXT-207) that are no
longer reproducible.