PXT-179 Add a if_exists parameter to create/add APIs
This commit adds a new parameter - if_exists - for the
add_columns, add_column, and add_computed_column APIs
which can be used to handle scenarios where a column
name may already exist. It can direct the API to not
add the column and return, or raise it as an error,
or replace the existing column with a new one.

Also fixed these APIs to raise an error when used
with a snapshot, instead of an assertion failure later
in the code.

Testing: added new testcases for the new parameter.