PXT-373 - Add if_not_exists parameter to drop_* APIs
This commit adds a if_not_exists parameter to the
drop_embedding_index and drop_index API. It can be used
to direct the API to either raise an error or do nothing
and return when an index does not exist. This only
applies when index name is specified.

Testing: adjusted existing callers and added coverage
for the new parameter.