PXT-388: Change the SQL Engine isolation level to 'REPEATABLE READ'
This commit changes the SQL Engine default isolation level
from AUTOCOMMIT to REPEATABLE READ. This exposed two bugs:

1. batch_update() API hangs when if_not_exists='insert'

RCA: In the "insert if not exists" mode, batch_update() translates
to two steps: a) update existing rows in the batch to new values,
and b) insert non-existing rows in the batch. The code was
using two different connection for these two steps. This resulted
in a self-deadlock when updating the table metadata row that is
common for both steps.

This commit fixes the code to use the same connection for both steps.

2. make_timestamp() resulted in sqlalchemy.exc.InternalError and
transcation abort.

RCA: make_timestamp() UDF to SQL conversion was producting wrong
SQL query. It cast a value to sql.Double which generates "double"
which is not a valid datatype in Postgres and results in an error.
Even though the error is seen irrespective of isolation level,
for repeated read the connection is lost and xact aborted.

This commit fixes the code to use sql.Float instead.