Support JSON Schemas for validating `JsonType` columns (#380)
Allows a `JsonType` to be qualified with a JSON Schema. In particular,
since every Pydantic model can be exported as a JSON Schema, this allows
Pydantic models to be used for validation. (However, custom Python
validators in the model will not be applied; only validators that are
representable in a JSON Schema are currently applied.)