PXT-371: add table context to ColumnRef
A column in the base table can be referenced by the table
or a view/snapshot derived from it. This commit adds a table
context property (in-memory only) to the ColumnRef expression
to differentiate the context in which the column is being
used/referenced. This is currently needed to use the right
embedding index.