Sample Apps: AI-Based Trading Chrome Extension (#428)
This is the open source version of my chrome extension using Pixeltable
and Claude Sonnet to take a screenshot of the active tab on Chrome,
store it in Pixeltable and leverages Pixeltable to orchestrate the LLM
calls and parsing of the outputs in order to deliver a JSON to the
frontend that can be displayed as a Chrome Extension.

It's working consistently and reliable. The experience is different
depending on the screenshot being taken. I optimized it for use with
Robinhood Legend but I showcase a screenshot with Yahoo. It'll work on
any charts it all depends on the information available on the charts and
what indicators have been displayed and configured.