PXT-373 - Add if_not_exists parameter to drop_* APIs (#423)
This commit adds a new parameter - if_not_exists - to the drop_dir,
drop_table,
drop_embedding_index, drop_index and drop_column APIs. This can be used
to
direct the API to either raise an error or do nothing and return when a
dir/table/index/column
does not exist.

Testing: added new test for the new parameter.