Tool calling (#416)
Adds a method `pxt.tools()` that allows a collection of UDFs to be cast
as LLM tools.

The output of `pxt.tools()` is an instance of a Pydantic model
`pxt.func.Tools` that can be invoked or converted to a dict.

This PR also upgrades `FunctionCall` to allow Pydantic model instances
as parameters. More generally, a Pydantic model instance can be cast as
a Literal, which will cause it to be converted to a dict via
`my_model.model_dump()`.