[SHOW] PXT-0416: Enable variable parameter in query limit() clause (#454)
Accomplishes the feature noted in the Jira ticket.

There are no tests yet, this PR is for feedback.

There seems to be a problem with the substitute() method in expr.py. The
problem is noted in comments in the method(). I will pend this change
until a decision is made about the operation of this method.

---------

Co-authored-by: jpeterson <jpeterson@pixeltable.com>