PXT-279: Refactor import methods, add functionality. (#489)
Modify type inference to include the concept of a type without a
supertype.
This concept is then used to short-circuit long running type inference
operations.
Add parquet files for testing.
Accept more types in some of the formats.
Map imported column names to valid pixeltable identifiers if needed.
Map data to pixeltable column names if needed.
All import methods now support schema_overrides and primary_key
parameters.
Report an error if any schema_overrids columns are not in the data
source
Report an error if any primary_key columns are not in the data source
Report an error when attempting to import rows with invalid column
names.