Fail gracefully when a backwards incompatibility is detected in a computed column UDF call (#494)
Improves error handling in the case where a computed column references a
UDF, but the UDF is no longer compatible with its definition in the
column, for one of the following reasons:
- The referenced UDF no longer exists (because it has been removed or
renamed from the code)
- The UDF signature(s) are no longer compatible with the call pattern in
the computed column
- The UDF return type is no longer compatible with the column's col_type

Previously, a table containing such a UDF would fail to load. Now, the
table will load and generate a warning when it is loaded. It can be
queried normally, and standard table operations (such as adding and
removing columns) can be carried out normally, but attempting to insert
data will result in an error.