Fix bug in JsonPath serialization (#532)
Fixes a bug in JsonPath serialization involving expressions with a
relative path root. Previously, such expressions inappropriately
serialized an `ObjectRef` instance, which could not be deserialized. The
new logic skips the `ObjectRef`, serializes the `JsonPath` as a relative
path, and re-binds it after deserialization.

Also changes `TableVersion` initialization to force column metadata to
load when the `TableVersion` is constructed. Previously, metadata
loading happened only on a call to `get_table`. But `DataFrame`
materialization from a dict (among other things) bypasses that code
path.