Bugfix related to chained tool calling (#534)
Fixes a bug in `JsonMapper` evaluation that caused relative expressions
to fail to bind when they appeared in the anchor of a `JsonPath`.

Also introduces a _workaround_ (not a proper fix) for a bug that causes
`JsonMapper` evaluation to fail when the target expression involves an
async UDF. The workaround forces the async UDF to be evaluated
synchronously, which will give the correct output, but introduces a
performance cliff when large numbers of rows are involved.