Update dev versions of many libraries used by Pixeltable (#666)
Updates many of the dependencies used in our dev environment including
torch, transformers, spacy, and many of our inference provider SDKs.

This is a necessary step towards being able to fully test in Python 3.13
as well as ensuring we are correctly supporting the latest versions of
our integrations.

Only the dev environment is affected; this will not affect
pypi/pyproject version compatibility.

The spacy upgrade required a change to a unit test, due to slightly
different output from the newer model.