Use a JSON field xxx_cellmd in place of xxx_errortype and xxx_errormsg (#685)
Add a new JSON field _cellmd which replaces _errortype and _errormsg
It contains two keys "errormsg" and "errortype" which take on the values
of the replaced columns
There are no changes to the metadata
All records in all user tables and views are rewritten in SQL to
consolidate values in the original columns into the new one.
The JSON keys "errormsg" and "errortype" are accessible via the
table.column.(errormsg/errortype) mechanism.
The full new field is accessible via table.column.cellmd