Extend FrameIterator to output all available frame attributes
This behavior is controlled by kwarg all_frame_attrs and disabled by default
to preserve compatibility with existing views created with this iterator.

This is undesirable (by default it should be on) and we should change that
in a follow-up, once we settle on a good approach to iterator output
schema evolution.