PXT-671: fixes to RateLimitsScheduler (#741)
The current rate limits scheduler discards the header-provided resource
utilization information when it receives 429 (rate limit error)
exceptions.

Tests with an Openai account that has standard (low) rate limits of 200k
TPM confirms that with the changes, the scheduler correctly holds off on
submitting requests after a 429 error until the indicated reset time,
even when subsequent (out-of-order) responses indicate that more
resources are available. The results is far fewer retries.