Group local media file operations into a MediaStore or TempStore class (#748)
The MediaStore interface is used to write, move, and delete files
containing media data persisted in a local filesystem directory.
The TempStore interface provides similar functionality for temporary
files and directories.
The TempStore acts as a stateless interface to the Env tmp_dir.

Both classes provide some interfaces for summarizing the contents of the
stores. These methods are only intended to be used with Pixeltable unit
tests in specific circumstances.