Update many error messages for consistency (#869)
Our error message formatting and verbiage is all over the place, and
it's been bugging me. On the plane I went through and audited many error
messages for consistency (but not all: we have a lot of them!)

I suggest standardizing on the following conventions, though I'm less
invested in the specific conventions than that there be a consistent
standard.
- Entity names and string literals are enclosed in single quotes.
```
Column 'my_col' does not have an index
```
- Code elements (parameter names, type literals, code snippets, etc) are
enclosed in backticks.
```
`on` expects an expression of type `Bool`
```
- However, a single literal that appears after a colon at the end of a
message is unquoted.
```
Unknown column: my_col
```

Comments welcome.