[PXT-822, PXT-674] Fix for querying snapshots of tables with unstored columns (#895)
Fixes a bug involving snapshots of tables with unstored columns;
previously, attempting to query them would trigger an `AssertionError`.
The issue was that column `value_expr`s in snapshot table versions were
not being retargeted.

The bug would also manifest in snapshots of views with explicit column
selections, as well as (data-sharing) replicas of views with explicit
column selections, both of which rely on snapshots over unstored
columns.

Fixes another (less critical) bug in the process, in which the
`str(computed_with)` display of a snapshot would use the incorrect
column names, in the event that the columns had been renamed in a later
version of the table. (See: PXT-674)