Manage `additional_md` from Catalog, rather than TableVersion (#913)
`additional_md` can't be managed from TableVersion because named
snapshots don't have TableVersion instances.

With this change, it's now managed directly from `Catalog`. In addition,
`additional_md` is moved to a separate column in the metadata tables
(rather than a field of `md`) so that updates to ordinary,
version-specific metadata in `md` don't interfere with it.

Also included is a working unit test for push/pull. (Hooray!)

The existing `md.additional_md` fields are kept in place for now so that
Cloud can continue to use them until a migration can be orchestrated.