`create_store_tbls: bool` option in Catalog.create_replica() (#916)
Pixeltable cloud needs to be able to create replica metadata instances
without actually creating the underlying store table.

This commit adds a `create_store_tbls` option to
`Catalog.create_replica()`. We need the option to be selective on a
per-table basis, since in the future we'll have both hosted tables
(which do have underlying store tables) and replicas (which don't)
living side by side, potentially in the same catalog.

There are various other ways to achieve this (e.g., have a flag in table
metadata that controls whether it has an underlying store table; or have
a flag in the environment that disables store tables specifically for
replicas). This PR seems like the most straightforward solution, but we
could consider those alternatives also.