Streamline Makefile (#937)
No one seems to be using or maintaining the Windows hooks in our
Makefile, and they're already broken in several places. This backs them
out so that the Makefile is once again Linux/MacOS only, which
streamlines the Makefile and makes it easier to maintain. If Windows
support becomes important again we'll need to redo it anyway.

(This won't affect any customers using Pixeltable on Windows.)