Rename `DataFrame` to `Query` and `DataFrameResultSet` to `ResultSet` (#902)
I tried to be as thorough as I could, renaming local variables and doc
references in addition to classes.

It wasn't entirely trivial, since `df` can also (and often does) mean
"Pandas DataFrame", and also appears as a substring of `udf`, `pdf`, and
`soundfile`. But I did my best to rename carefully.