Support audio and video embedding indices (#990)
Adds support for embedding indices of type audio and video.

It's now possible to *index* an audio or video column, but it's still
not possible to do *lookups* by passing an audio or video literal to
similarity(), because we don't have a way to construct audio or video
literals currently (they just look like strings); I've left this as a
todo for now.